requirements.txt
Flask==2.3.2
gunicorn==20.1.0
APScheduler==3.10.1
pandas==2.0.3
reportlab==4.0.4
twilio==8.5.0
sendgrid==6.10.0
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
Flask-Moment==1.0.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
# New additions for Phase 3
pytest==7.4.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
